+++ anuga/utilities/cg.c
@@ -11,13 +11,14 @@
 //
 // Note Padarn 5/12/12: I have tried a few optimization modifications which
 // didn't seem to save any time:
-// -- conversion of the long arrays to int arrays (to save memory passing time)
+// -- conversion of the int64_t arrays to int64_t arrays (to save memory passing time)
 // -- taking advantage of the symmetric quality of the matrix A to reduce the zAx loop
 // -- specifying different 'chunk' sizes for the openmp loops
 // -- using blas instead of own openmp loops
 	
 #include "math.h"
 #include "stdio.h"
+#include <stdint.h>
 
 #if defined(__APPLE__)
    // clang doesn't have openmp
@@ -28,14 +29,14 @@
 
 
 // Dot product of two double vectors: a.b
-// @input N: int length of vectors a and b
+// @input N: int64_t length of vectors a and b
 //        a: first vector of doubles
 //        b: second vector of double
 // @return: double result of a.b 
-double cg_ddot( int N, double *a, double *b)
+double cg_ddot( int64_t N, double *a, double *b)
 {
   double ret = 0;
-  int i;
+  int64_t i;
   #pragma omp parallel for private(i) reduction(+:ret)
   for(i=0;i<N;i++)
   {
@@ -46,12 +47,12 @@ double cg_ddot( int N, double *a, double *b)
 }
 
 // In place multiplication of a double vector x by constant a: a*x
-// @input N: int length of vector x
+// @input N: int64_t length of vector x
 //        a: double scalar to multiply by
 //        x: double vector to scale
-void cg_dscal(int N, double a, double *x)
+void cg_dscal(int64_t N, double a, double *x)
 {
-  int i;
+  int64_t i;
   #pragma omp parallel for private(i)
   for(i=0;i<N;i++)
   {
@@ -61,12 +62,12 @@ void cg_dscal(int N, double a, double *x)
 }
 
 // Copy of one vector to another - memory already allocated: y=x
-// @input N: int length of vectors x and y
+// @input N: int64_t length of vectors x and y
 //        x: double vector to make copy of
 //        y: double vector to copy into
-void cg_dcopy( int N, double *x, double *y)
+void cg_dcopy( int64_t N, double *x, double *y)
 {
-  int i;
+  int64_t i;
   #pragma omp parallel for private(i)
   for(i=0;i<N;i++)
   {
@@ -75,13 +76,13 @@ void cg_dcopy( int N, double *x, double *y)
 }
 
 // In place axpy operation: y = a*x + y
-// @input N: int length of vectors x and y
+// @input N: int64_t length of vectors x and y
 //        a: double to multiply x by
 //        x: first double vector
 //        y: second double vector, stores result
-void cg_daxpy(int N, double a, double *x, double *y)
+void cg_daxpy(int64_t N, double a, double *x, double *y)
 {
-  int i;
+  int64_t i;
   #pragma omp parallel for private(i)
   for(i=0;i<N;i++)
   {
@@ -92,15 +93,15 @@ void cg_daxpy(int N, double a, double *x, double *y)
 // Sparse CSR matrix-vector product: z = A*x
 // @input z: double vector to store the result
 //        data: double vector with non-zero entries of A
-//        colind: long vector of column indicies of non-zero entries of A
-//        row_ptr: long vector giving index of rows for non-zero entires of A
+//        colind: int64_t vector of column indicies of non-zero entries of A
+//        row_ptr: int64_t vector giving index of rows for non-zero entires of A
 //        x: double vector to be multiplied
 //        M: length of vector x
-void cg_zAx(double * z, double * data, long * colind, long * row_ptr, double * x, int M){
+void cg_zAx(double * z, double * data, int64_t * colind, int64_t * row_ptr, double * x, int64_t M){
 
   
   
-  long i, j, ckey;
+  int64_t i, j, ckey;
 
 
      
@@ -121,10 +122,10 @@ void cg_zAx(double * z, double * data, long * colind, long * row_ptr, double * x
 //        D: double vector of diagonal matrix
 //        x: double vector to be multiplied
 //        M: length of vector x
-void cg_zDx(double * z, double * D, double * x, int M){
+void cg_zDx(double * z, double * D, double * x, int64_t M){
 
   
-  long i, j, ckey;
+  int64_t i, j, ckey;
    
     #pragma omp parallel for private(ckey,j,i)
     for (i=0; i<M; i++){
@@ -139,10 +140,10 @@ void cg_zDx(double * z, double * D, double * x, int M){
 //        D: double vector of diagonal matrix
 //        x: double vector to be multiplied
 //        M: length of vector x
-void cg_zDinx(double * z, double * D, double * x, int M){
+void cg_zDinx(double * z, double * D, double * x, int64_t M){
 
   
-  long i, j, ckey;
+  int64_t i, j, ckey;
    
     #pragma omp parallel for private(ckey,j,i)
     for (i=0; i<M; i++){
@@ -158,14 +159,14 @@ void cg_zDinx(double * z, double * D, double * x, int M){
 // @input z: double vector to store the result
 //        a: double to scale matrix-vector product by
 //        data: double vector with non-zero entries of A
-//        colind: long vector of column indicies of non-zero entries of A
-//        row_ptr: long vector giving index of rows for non-zero entires of A
+//        colind: int64_t vector of column indicies of non-zero entries of A
+//        row_ptr: int64_t vector giving index of rows for non-zero entires of A
 //        x: double vector to be multiplied
 //        y: double vector to add
 //        M: length of vector x
-void cg_zaAxpy(double * z, double a, double * data, long * colind, long * row_ptr, double * x,
-      double * y,int M){
-  long i, j, ckey;
+void cg_zaAxpy(double * z, double a, double * data, int64_t * colind, int64_t * row_ptr, double * x,
+      double * y,int64_t M){
+  int64_t i, j, ckey;
   #pragma omp parallel for private(ckey,j,i)
     for (i=0; i<M; i++ ){
       z[i]=y[i];
@@ -182,19 +183,19 @@ void cg_zaAxpy(double * z, double a, double * data, long * colind, long * row_pt
 // Jacobi preconditioner for matrix, A, and right hand side, b. Mutiplies each row
 // by one divided by the diagonal element of the matrix. If the diagonal 
 // element is zero, does nothing (should nnot occur)
-//        colind: long vector of column indicies of non-zero entries of A
-//        row_ptr: long vector giving index of rows for non-zero entires of A
+//        colind: int64_t vector of column indicies of non-zero entries of A
+//        row_ptr: int64_t vector giving index of rows for non-zero entires of A
 //        b: double vector specifying right hand side of equation to solve
 //        M: length of vector b
 
-int _jacobi_precon_c(double* data, 
-                long* colind,
-                long* row_ptr,
+int64_t _jacobi_precon_c(double* data, 
+                int64_t* colind,
+                int64_t* row_ptr,
                 double * precon,
-                int M){
+                int64_t M){
 
 
-  long i, j, k, ckey;
+  int64_t i, j, k, ckey;
   double diag;
 
 
@@ -220,25 +221,25 @@ int _jacobi_precon_c(double* data,
 
 // Conjugate gradient solve Ax = b for x, A given in Sparse CSR format
 // @input data: double vector with non-zero entries of A
-//        colind: long vector of column indicies of non-zero entries of A
-//        row_ptr: long vector giving index of rows for non-zero entires of A
+//        colind: int64_t vector of column indicies of non-zero entries of A
+//        row_ptr: int64_t vector giving index of rows for non-zero entires of A
 //        b: double vector specifying right hand side of equation to solve
 //        x: double vector with initial guess and to store result
 //        imax: maximum number of iterations
 //        tol: error tollerance for stopping criteria
 //        M: length of vectors x and b
 // @return: 0 on success  
-int _cg_solve_c(double* data, 
-                long* colind,
-                long* row_ptr,
+int64_t _cg_solve_c(double* data, 
+                int64_t* colind,
+                int64_t* row_ptr,
                 double * b,
                 double * x,
-                int imax,
+                int64_t imax,
                 double tol,
                 double a_tol,
-                int M){
+                int64_t M){
 
-  int i = 1;
+  int64_t i = 1;
   double alpha,rTr,rTrOld,bt,rTr0;
 
   double * d = malloc(sizeof(double)*M);
@@ -290,8 +291,8 @@ int _cg_solve_c(double* data,
 // Conjugate gradient solve Ax = b for x, A given in Sparse CSR format,
 // using a diagonal preconditioner M. 
 // @input data: double vector with non-zero entries of A
-//        colind: long vector of column indicies of non-zero entries of A
-//        row_ptr: long vector giving index of rows for non-zero entires of A
+//        colind: int64_t vector of column indicies of non-zero entries of A
+//        row_ptr: int64_t vector giving index of rows for non-zero entires of A
 //        b: double vector specifying right hand side of equation to solve
 //        x: double vector with initial guess and to store result
 //        imax: maximum number of iterations
@@ -299,18 +300,18 @@ int _cg_solve_c(double* data,
 //        M: length of vectors x and b
 //        precon: diagonal preconditioner given as vector
 // @return: 0 on success  
-int _cg_solve_c_precon(double* data, 
-                long* colind,
-                long* row_ptr,
+int64_t _cg_solve_c_precon(double* data, 
+                int64_t* colind,
+                int64_t* row_ptr,
                 double * b,
                 double * x,
-                int imax,
+                int64_t imax,
                 double tol,
                 double a_tol,
-                int M,
+                int64_t M,
                 double * precon){
 
-  int i = 1;
+  int64_t i = 1;
   double alpha,rTr,rTrOld,bt,rTr0;
 
   double * d = malloc(sizeof(double)*M);
